Mudanças nos repositories para apenas um
