+++ app/src/main/java/br/com/boa50/kbingo/KbingoApplication.java
@@ -2,8 +2,7 @@ package br.com.boa50.kbingo;
 
 import javax.inject.Inject;
 
-import br.com.boa50.kbingo.data.source.LetrasRepository;
-import br.com.boa50.kbingo.data.source.PedrasRepository;
+import br.com.boa50.kbingo.data.AppRepository;
 import br.com.boa50.kbingo.di.DaggerAppComponent;
 import dagger.android.AndroidInjector;
 import dagger.android.DaggerApplication;
@@ -15,10 +14,7 @@ import dagger.android.DaggerApplication;
 public class KbingoApplication extends DaggerApplication {
 
     @Inject
-    PedrasRepository pedrasRepository;
-
-    @Inject
-    LetrasRepository letrasRepository;
+    AppRepository appRepository;
 
     @Override
     protected AndroidInjector<? extends DaggerApplication> applicationInjector() {
