+++ app/src/main/java/br/com/boa50/kbingo/data/AppDatabase.java
@@ -1,0 +1,4 @@
+package br.com.boa50.kbingo.data;
+
+public class AppDatabase {
+}
