+++ app/src/main/java/br/com/boa50/kbingo/data/AppRepository.java
@@ -1,4 +1,6 @@
-package br.com.boa50.kbingo.data.source;
+package br.com.boa50.kbingo.data;
+
+import com.google.common.collect.Lists;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,30 +8,36 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import br.com.boa50.kbingo.data.Letra;
-import br.com.boa50.kbingo.data.Pedra;
+import br.com.boa50.kbingo.data.entity.Letra;
+import br.com.boa50.kbingo.data.entity.Pedra;
 import br.com.boa50.kbingo.util.StringUtils;
 import io.reactivex.Flowable;
 
+import static br.com.boa50.kbingo.Constant.QTDE_LETRAS;
 import static br.com.boa50.kbingo.Constant.QTDE_PEDRAS_LETRA;
 
-/**
- * Created by boa50 on 3/24/18.
- */
-
 @Singleton
-public class PedrasRepository implements PedrasDataSource {
+public class AppRepository implements AppDataSource {
 
     @Inject
-    LetrasRepository letrasRepository;
+    AppRepository() {}
 
-    @Inject
-    PedrasRepository() {}
+    @Override
+    public List<Letra> getLetras() {
+        ArrayList<Letra> array =  Lists.newArrayList(
+                new Letra("1", "K", 0),
+                new Letra("2", "I", 1),
+                new Letra("3", "N", 2),
+                new Letra("4", "K", 3),
+                new Letra("5", "A", 4)
+        );
+
+        return array.subList(0, QTDE_LETRAS);
+    }
 
     @Override
     public Flowable<List<Pedra>> getPedras() {
-
-        List<Letra> letras = letrasRepository.getLetras();
+        List<Letra> letras = getLetras();
         List<Pedra> pedras = new ArrayList<>();
 
         for (int i = 1; i <= QTDE_PEDRAS_LETRA*letras.size(); i++) {
