+++ app/src/main/java/br/com/boa50/kbingo/data/dao/PedraDao.java
@@ -1,0 +1,4 @@
+package br.com.boa50.kbingo.data.dao;
+
+public class PedraDao {
+}
