--- app/src/main/java/br/com/boa50/kbingo/data/source/LetrasDataSource.java
@@ -1,14 +1,0 @@
-package br.com.boa50.kbingo.data.source;
-
-import java.util.List;
-
-import br.com.boa50.kbingo.data.Letra;
-
-/**
- * Created by boa50 on 4/1/18.
- */
-
-public interface LetrasDataSource {
-
-    List<Letra> getLetras();
-}
